data-scraper/requirements.txt
psycopg2-binary==2.9.7
requests==2.31.0
boto3==1.34.0
python-dotenv==1.0.0